+++ core/commons/suite_execution_volume/suite_execution_volume_test.go
@@ -13,36 +13,20 @@ import (
 	"testing"
 )
 
-func TestGetSuiteExecutionDirectory(t *testing.T) {
+func TestGetEnclaveDirectory(t *testing.T) {
 	suiteExVolDirpath, err := ioutil.TempDir("", "")
 	assert.Nil(t, err)
 
 	testId := "someTest"
 
 	suiteExVol := NewSuiteExecutionVolume(suiteExVolDirpath)
-	testExDir, err := suiteExVol.GetEnclaveDirectory([]string{testId})
+	enclaveDir, err := suiteExVol.GetEnclaveDirectory([]string{testId})
 	assert.Nil(t, err)
 
 	expectedAbsDirpath := path.Join(suiteExVolDirpath, testId)
 	_, err = os.Stat(expectedAbsDirpath)
 	assert.Nil(t, err)
 
-	assert.Equal(t, expectedAbsDirpath, testExDir.absoluteDirpath)
-	assert.Equal(t, testId, testExDir.dirpathRelativeToVolRoot)
-}
-
-func TestGetArtifactCache(t *testing.T) {
-	suiteExVolDirpath, err := ioutil.TempDir("", "")
-	assert.Nil(t, err)
-
-	suiteExVol := NewSuiteExecutionVolume(suiteExVolDirpath)
-	artifactCache, err := suiteExVol.GetArtifactCache()
-	assert.Nil(t, err)
-
-	expectedAbsDirpath := path.Join(suiteExVolDirpath, artifactCacheDirname)
-	_, err = os.Stat(expectedAbsDirpath)
-	assert.Nil(t, err)
-
-	assert.Equal(t, expectedAbsDirpath, artifactCache.absoluteDirpath)
-	assert.Equal(t, artifactCacheDirname, artifactCache.dirpathRelativeToVolRoot)
+	assert.Equal(t, expectedAbsDirpath, enclaveDir.absoluteDirpath)
+	assert.Equal(t, testId, enclaveDir.dirpathRelativeToVolRoot)
 }